task: add Task::from_closure with more efficient boxing
This implementation of Task::from_closure (and spawn_async())
avoids the need for doubly-nested Boxes by giving the (sized)
closure type an explicit name as a generic parameter.

Since the passed-in closure type has a known size at compile
time, we can properly pass it to a newly-spawned Task via a
single pointer.

This has the downside of requiring these functions to monomorphize
to every single closure type used, however.